Add Control-C handler to manage interrupt signal
The code now includes a Control-C handler to manage the Control-C (interrupt) signal. This is done by defining a `CtrlHandler` function that handles the `CTRL_C_EVENT` by printing a message and returning `TRUE` to indicate the event was handled. For other control events, it returns `FALSE`. The `SetConsoleCtrlHandler` function is then used in the `main` function to set this handler. If setting the handler fails, an error message is printed and the program exits with a return code of `-1`.